chore(dw): clean up
